% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circle.R
\name{circle}
\alias{circle}
\title{Encircle a set of 2D points using quadratic Bezier curves}
\usage{
circle(data, r = 0.3, n_b = 5, dupl_first = FALSE)
}
\arguments{
\item{data}{A matrix or a data frame with two columns corresponding to the
x and y coordinates of the points.}

\item{r}{The 'radius' of the encircling shape. Used as the distance from the
points of the convex hull to the points used to create the bezier curves.}

\item{n_b}{The number of points used near each corner of the encircling
shape. Can be used to control the smoothness of the shape.}

\item{dupl_first}{If \code{TRUE}, the first point is added to the end of the
list. Useful for plotting the shapes using e.g. geom_path in ggplot2.
Defaults to \code{FALSE}.}
}
\value{
A matrix with points of the encircling shape as rows.
}
\description{
Find a smooth convex shape that encircles the points given as input.
}
\examples{
\donttest{
# A minimal example
mat <- matrix(rnorm(10), 5)
circ <- circle(mat)
}

}
