% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circle.R
\name{circle}
\alias{circle}
\title{Encircle a set of 2D points}
\usage{
circle(data, r = 0.5, s = 8, method = "b", closed = FALSE)
}
\arguments{
\item{data}{Either a matrix or a data frame with two real-valued columns
or a complex-valued vector.}

\item{r}{The distance from the data to the encircling shape.
Must be non-negative. Defaults to 0.5.}

\item{s}{A smoothness parameter. Must be greater than 2. For bezier curves,
this equals to the number of points around each point on the convex hull and
for method = "c", scaled to yield similar results. Defaults to 8.}

\item{method}{Either "b" for a quadratic Bezier curve or "c" for a
semicircle around each point of the hull. Defaults to "b".}

\item{closed}{If \code{TRUE}, the first point is added to the end of the
list. Useful for plotting the shapes using e.g. geom_path in ggplot2.
Defaults to \code{FALSE}.}
}
\value{
A matrix with points of the encircling shape as rows.
}
\description{
Find a smooth convex shape that encircles the points given as input using
either Bezier curves or semicircles around each corner.
}
\examples{
\donttest{
# A minimal example
data <- data.frame(x = rnorm(5), y = rnorm(5))
circ <- circle(data, closed = TRUE)
ggplot(circ) + geom_path(aes(x = x, y = y))
}


}
